% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{car_sales}
\alias{car_sales}
\alias{carsales}
\title{Car sale prices}
\source{
\href{https://www.kaggle.com/datasets/nehalbirla/vehicle-dataset-from-cardekho}{Kaggle datasets}
}
\value{
\item{car_sales}{a tibble}
}
\description{
These data were taken from \href{https://www.kaggle.com/datasets/nehalbirla/vehicle-dataset-from-cardekho}{Kaggle datasets}
and consists of sale prices, in INR (Indian Rupee), of cars.
}
\details{
The columns are:
\itemize{
\item \code{log_sale_price}: a double column (log, base 10).
\item \code{year}: a double column.
\item \code{km_driven}: a double column.
\item \code{fuel}: a factor column with 2 levels.
\item \code{seller_type}: a factor column with 3 levels: \code{'dealer'}, \code{'individual'}, and \code{'trustmark_dealer'}.
\item \code{transmission}: a factor column with 2 levels: \code{'automatic'} and \code{'manual'}.
\item \code{owner}: a factor column with 5 levels.
\item \code{mileage}: a double column.
\item \code{engine}: a double column.
\item \code{max_power}: a double column.
\item \code{seats}: a double column.
\item \code{keyword_abs} - \code{keyword_other}: binary indicators for keywords derived from
the original car name field (e.g. "Maruti Swift Dzire VDI").
}

Some rows were removed due to excessive missingness or parsing difficulties.
}
\examples{
data(car_sales)
str(car_sales)
}
\keyword{datasets}
