\docType{package}
\name{carsales-package}
\alias{carsales}
\alias{carsales-package}
\title{Car sales analysis for personal use}
\description{
  Car sales analysis for personal use
}
\details{
  \tabular{ll}{ Package: \tab carsales\cr Type: \tab
  Package\cr Version: \tab 0.2-1\cr Date: \tab 2013-10\cr
  License: \tab GPL 3\cr }
}
\author{
  J-M
}

