\name{getInfo}
\alias{getInfo}
\title{Extract car sales information from a list of XMLNode}
\usage{
  getInfo(infoHeadNodes)
}
\arguments{
  \item{infoHeadNodes}{a list of objects of class XMLNode}
}
\value{
  a named character vector
}
\description{
  Extract car sales information from a list of XMLNode
}

