% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carsales-package.r
\docType{package}
\name{carsales-package}
\alias{carsales}
\alias{carsales-package}
\title{Car sales analysis for personal use}
\description{
Car sales analysis for personal use
}
\details{
\tabular{ll}{
Package: \tab carsales\cr
Type: \tab Package\cr
Version: \tab 0.3-1\cr
Date: \tab 2015-11\cr
License: \tab GPL 3\cr
}
}
\author{
J-M
}

